big bugfix!! (see comments)
Failure had been happening because the API returned 200 OK for full game
responses, and custom 299 for instances such as "Summoner Not In-Game."
However, in the front end, Axios was reading 299 as 200, so errors were
being treated as successes. I fixed this by having caught errors return
as 200 with body {"err": ...errror} and updating the front end
accordingly.